make import_status2019 part of scenario_path dataset
