+++ src/egon/data/airflow/dags/pipeline.py
@@ -98,7 +98,7 @@ from egon.data.datasets.vg250_mv_grid_districts import Vg250MvGridDistricts
 from egon.data.datasets.zensus import ZensusMiscellaneous, ZensusPopulation
 from egon.data.datasets.zensus_mv_grid_districts import ZensusMvGridDistricts
 from egon.data.datasets.zensus_vg250 import ZensusVg250
-from egon.data.datasets.scenario_path.import_status2019 import Import_Status2019
+from egon.data.datasets.scenario_path import CreateIntermediateScenarios
 
 # Set number of threads used by numpy and pandas
 set_numexpr_threads()
@@ -673,8 +673,8 @@ with airflow.DAG(
         ]
     )
 
-    # import scenario status2019 from backup
-    import_status2019 = Import_Status2019(
+    # Create intermediate scenarios based on status2019 and eGon100RE
+    create_intemediate_scenarios = CreateIntermediateScenarios(
         dependencies=[
             storage_etrago,
             hts_etrago_table,
